We might be at 0.1!
RAG works for now. We need to include all the other endpoints now. The repo map, the git diff. Also, we need to fix the way we're displaying this context in the frontend as well. We need to show line numbers, not the start and end columns. Although that's almost all the information needed.